refactor: extract implement phase and PR helpers from fix-issue/workflow.ts
